backend/pyproject.toml
[project]
name = "forecaster-enterprise-backend"
version = "0.1.0"
description = "Forecaster Enterprise Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.11.7",
    "email-validator>=2.1.0", # Required for EmailStr validation
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pwdlib[argon2]>=0.3.0", # Modern password hashing (recommended by FastAPI, uses Argon2)
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.6",
    "asyncpg>=0.29.0", # Async PostgreSQL adapter
    "greenlet>=3.0.0", # Required for async SQLAlchemy
    "alembic>=1.13.0", # Database migrations
    "slowapi>=0.1.9", # Rate limiting
    "pandas>=2.0.0", # Data manipulation
    "numpy>=1.24.0", # Numerical operations
    "chronos-forecasting>=2.1.0", # Chronos-2 forecasting model
    "psycopg2-binary>=2.9.11",
    "darts>=0.39.0",
    "kaggle>=1.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.25.0",
    "aiosqlite>=0.19.0",  # For async SQLite testing
    "ruff>=0.12.12",
    "mypy>=1.17.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "UP", "I", "B006"]
ignore = ["E501"]
fixable = ["ALL"]
exclude = ["*.ipynb"]

